Agregados tests, cambie precio a double en algobay, implemente que al agregar varios cupones vale solo el mayor.

+++ TP1J/src/src/AlgoBay.java
@@ -13,7 +13,7 @@ public class AlgoBay {
 		return catalogo.getCantidadDeProductos();
 	}
 
-	public Producto agregarProductoConPrecio(String nombre, int precio) {
+	public Producto agregarProductoConPrecio(String nombre, double precio) {
 		return catalogo.agregarProductoConPrecio(nombre,precio);
 	}
